chore: mv util functions out of strandedness and into util
